Add Vundle package manager for Vim
Vundle is a submodule in the `files/vim/bundle/` directory, in which
bundles are to be stored. (A git ignore file has been added to exclude
other bundles from the repository.) Vim can be made to install or update
the bundles specified in `vimrc` with the `:BundleInstall` command, or
`vim +BundleInstall! +qall` from the command line.